@@ -5,7 +5,7 @@ author: StoneHui <shichaohui1230@gmail.com>
 homepage: https://github.com/shichaohui/gesture_zoom_box
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
